Create connection.php
this code if for the connection between the index.php and the "login.php" (not coded yet).

please name your database "DB1" when create the database that includes the username table on your phpmyadmin.

To confirm that the connection is workin, check for  a connection successful status on your localhost page.